refactor: remove global test server process, use per-file process management
